Fortran Analyzer: TOOLS/f2py/fparser (r24. See http://code.google.com/p/f2py/)
TOOLS/fatest.py fp/*.F gwd/*.F slatsm/*.F subs.*.F
works fine (but still a test version). But they are still a kind of
demo version. We need to do data analysis on it.

All required fortran codes are slightly modified so that
the analyzer can analyze lm7K.

Thus we need the above modification in fortrn codes so as to satisfy:
 (1) these direvtives(#if blocks) are well nested in the fortran code.
 (2) No comment lines within format.--->need check more for analyzer.
 (3) Parameter sentences choosen by #if blocks are not allowed.
 (4) No assign statement.(even in previous version).
I had no problem to modify lm7K to satisify them.

fparser is modified sligtly.
1. _is_fix_comment = lambda line: line and line[0] in '*cC!#'
  at raedfortran.py L36 (# is added).
2. ##       if not line:
   ##           return self.get_single_line()
   are commented out at the end of get_single_line

Now fpp directives (#if and so) are recognized just as comment lines for analyzer.
Now PLATFORM=gfortran.fpp is not necessary.