Corrected TOOLS/f_calltree.py, which is called from CallCaller.sh. It is useful, but it may still contain a problem. (there is a possibity that some function calls are missing from the callcaller.dat).
The item.line did not reproduced original lines;
in cases, it is replaced by F2PY_foobar. So, it may be better to
use item.strline and item strlinemap. 

But function call (not subroutine call) can be nested.
Then F2PY_something must have a nested structure.
Can we analyze it?