lm72-gga files are taken. this versio esstially pass make check.
---------------------------
It fails for co c crn, but differences are small.
I think it is because the new version of exc calculator
correct something (positive rho correction?).

We need to modify locpt2 and smvxc2 as lm72 to treat GGA.
(NOTE: lxcf swich has difference; 100's digit for
core ---> 10000's digit for core. See explanation at the top of vxcsp.).

vxc0sp and makrm0 calls new versio of evxcv and evxcp. 
I think may be not need to modify them.