zhev.F and m_rdctl.F
They are moved back to previous ones (scaling in zhev is recovered).
default value of HAM_STABILIZE=1d-8. A little safer, though slight
poorness in total energy(I think it it not a matter).
Try HAM_STABILIZE 1d-10---> but it may cause a stop because of poor
linear dependency (NULL vector can be included in eigenvector).