make check OK. gw_lmfh OK for newaniso. But 1.samples not organized well. 2. EIBZdeveloping
Look into explanation for eibz4x0() in fpgw/gwsrc/switch.F.
Current implementation of EIBZ mode is still developing.
(only timereversal=.false. works; it is for development and very
 inefficient for symmetrization at the bottom of x0kf_v4h.)

So, TODO are
(1) EIBZ mode. probably it is better to start from q <---> -q symmetric q points.
(2) gwsc is not for newaniso mode.
(3) Tests for anisotropic case.