Temporary. Mainly eibzmode (see switch.F). It works(but slow).
This is a temporary version of EIBZ for gw_lmfh.
Symmetrization at the end of x0kf_v4h.F works OK (for test_si_gw_lmfh
case). But it is slow; I have to improve the mechanism on rotMPB.
(this can be given as a simple index changes with phase
multiplication).