Fix Segmenation fault due to stacksize error.
 ulimit -s unlimited works. But it can be inconvenient. 
Thus I fixed some locations of codes, not to require stack but to
allocate explicitely. For larger calculations, we may have some other
parts where we further need to fix.

When we call 
   call foobar( ,dreal(arrayx),...
, we have automatic array for dreal(arrayx). This consumes stack.
(the automatic array can be decleared at the begining of a rouitne).

When ulimit -a shows small stacksize, it can stop with SegFault.