Bugfix. We need QpGcut_psi2 = (2d0+1d-2)*QpGcut_psi + 2d0*pi/alat*dQpG in rdata4gw_v2.m.F.
NOTE: ngc2 which is to expand the IPW parts of \phi \times \phi.
In principle, QpGcut_psi2 = 2d0*QpGcut_psi. 

But we need to add 2d0*pi/alat*dQpG so that all G of
|Q0P+G|<= 2*QpGcut_psi are taken for all Q0P.
(the IPW set of G vectors determined at the Gamma point 
are used at other Q0P points.)