mkqg is changed (setting of Q0P). This does not affects to highy symmetric case.
For full tensor of dielectric function at Gamma, we first prepare six
independent directions and remove some of them when they are
lineraly-dependent. Because of the q-dependent of the dielectric
functions, (we may take q \to zero limit, however, the q-dependence may
be too large to take such a limit).

As six q vectors, we take 1/2*b1, 1/2*b2, 1/2*b3, and 
their sum or diffrences (smaller ones). This can be a reasonable choice.