make all OK! with tEST version of PWMODE=1 and 2 for gwsc.
We have
  q + igv2@gamma --> q + ((qs-q) + igv2@gamma),
where qs-q is nlatout.
(in other words,  igv2@gamma --> igv2=igv2@gamma +nlatout)
For given q, q=qs-q is on of the shortest vectors.

Our new igv2 contains G vectors (qs-q) + igv2@gamma.

call shortn3 ---------------
(
q   = pin
qs  = pin + nlatout
qs-q = nlatout
----------------------------

As for gwsc for pwmode=1,2,
we need to skip symmetry operation in eibzgen (in mptauof)
when q in on the boundary of BZ.

As for lmf (bndfp.F), we add modification to add nlatout, so that
all G vector is defined at Gamma point.