bugfix (probably, only it had affected to pdos mode for nspin=2). paralell mode mof job_band_nspin and job_tdos
See rdsigm2.F. and hqpe.sc.F Now sigm file contains isp.
Bug was in the possibility that we read
sigm for given q but with wrong spin index.

job_band_nspin1, job_band_nspin2, and job_tdos are for k-paralell.
But job_pdos for k-paralell is not yet implemented.
(probably not so difficult